Move side menu to Angular (instantiationStatus {AuditInfo})
Issue-ID: VID-759

+++ vid-webpack-master/src/app/shared/components/auditInfoModal/auditInfoModal.component.ts
@@ -1,4 +1,4 @@
-import {Component, ViewChild} from '@angular/core';
+import {Component, Input, ViewChild} from '@angular/core';
 import {Subject} from 'rxjs/Subject';
 import {ModalDirective} from 'ngx-bootstrap'
 import {ModelInformationItem} from '../model-information/model-information.component';
@@ -20,7 +20,10 @@ import {FeatureFlagsService, Features} from "../../services/featureFlag/feature-
 export class AuditInfoModalComponent {
   static openModal: Subject<ServiceInfoModel> = new Subject<ServiceInfoModel>();
   static openInstanceAuditInfoModal: Subject<{instanceId , type, model, instance}> = new Subject<{instanceId , type, model, instance}>();
+  modalIsOpen : boolean = false;
   @ViewChild('auditInfoModal', {static: false}) public auditInfoModal: ModalDirective;
+  @Input() isIframe : boolean = false;
+
   title: string = 'Service Instantiation Information';
   modelInfoItems: ModelInformationItem[] = [];
   serviceModel: ServiceModel;
@@ -45,21 +48,28 @@ export class AuditInfoModalComponent {
               private store: NgRedux<AppState>) {
     this.auditInfoModalComponentService = this._auditInfoModalComponentService;
     AuditInfoModalComponent.openModal.subscribe((jobData: ServiceInfoModel) => {
+      this.modalIsOpen = true;
       this.isALaCarteFlagOn = this.store.getState().global.flags['FLAG_A_LA_CARTE_AUDIT_INFO'];
       this.showMoreAuditInfoLink = _featureFlagsService.getFlagState(Features.FLAG_MORE_AUDIT_INFO_LINK_ON_AUDIT_INFO);
       this.initializeProperties();
       this.showVidStatus = true;
       if (jobData) {
         this.isAlaCarte = jobData.aLaCarte;
-        this.openAuditInfoModal(jobData);
-        _iframeService.addClassOpenModal(this.parentElementClassName);
+        this.openAuditInfoModal(jobData, this.isIframe);
+        if(this.isIframe){
+          _iframeService.addClassOpenModal(this.parentElementClassName);
+        }
         this.serviceModelName = jobData.serviceModelName ? jobData.serviceModelName : '';
         this.serviceModelId = jobData.serviceModelId;
         this.jobId = jobData.jobId;
         this.auditInfoModal.show();
+        this.modalIsOpen = true;
       } else {
-        _iframeService.removeClassCloseModal(this.parentElementClassName);
+        if(this.isIframe){
+          _iframeService.removeClassCloseModal(this.parentElementClassName);
+        }
         this.auditInfoModal.hide();
+        this.modalIsOpen = false;
       }
     });
 
@@ -82,6 +92,7 @@ export class AuditInfoModalComponent {
       this.modelInfoItems = this.auditInfoModalComponentService.getModelInfo(model, instance, instanceId);
       _iframeService.addClassOpenModal(this.parentElementClassName);
       this.auditInfoModal.show();
+      this.modalIsOpen = true;
     });
   }
 
@@ -98,14 +109,17 @@ export class AuditInfoModalComponent {
     this.isLoading = true;
   }
 
-  openAuditInfoModal(jobData: ServiceInfoModel): void {
+  openAuditInfoModal(jobData: ServiceInfoModel, isIframe ?: boolean): void {
     this.modelInfoItems = AuditInfoModalComponentService.createModelInformationItemsJob(jobData);
     this.initAuditInfoData(jobData);
     this.auditInfoModal.onHide.subscribe(()=>{
-      this._iframeService.removeClassCloseModal(this.parentElementClassName);
+      if(isIframe){
+        this._iframeService.removeClassCloseModal(this.parentElementClassName);
+      }
       this.initializeProperties();
     });
     this.auditInfoModal.show();
+    this.modalIsOpen = true;
   }
 
   initAuditInfoData(jobData: ServiceInfoModel) {
@@ -121,6 +135,7 @@ export class AuditInfoModalComponent {
     this._iframeService.removeClassCloseModal(this.parentElementClassName);
     this.initializeProperties();
     this.auditInfoModal.hide();
+    this.modalIsOpen = false;
   }
